+++ lib/stenotype/at_exit.rb
@@ -1,6 +1,8 @@
-require 'stenotype'
+# frozen_string_literal: true
 
 # :nocov:
+require "stenotype"
+
 at_exit do
   targets = Stenotype.config.targets
   targets.each(&:flush!)