+++ lib/stenotype/railtie.rb
@@ -18,6 +18,8 @@ module Stenotype
 
     config.stenotype = Stenotype.config
 
+    config.after_initialize { config.stenotype.targets.each(&:setup!) } if config.stenotype.rails.explicit_client_initialization
+
     if config.stenotype.rails.enable_action_controller_ext
       ActiveSupport.on_load(:action_controller) do
         Stenotype::ContextHandlers.register Stenotype::ContextHandlers::Rails::Controller