+++ spec/stenotype/event_spec.rb
@@ -60,23 +60,22 @@ RSpec.describe Stenotype::Event do
       end
     end
 
-    context 'when exception is raised' do
+    context "when exception is raised" do
       subject(:emit_event) { described_class.emit!("some_event", *{ key: :value }) }
 
       let(:logger_double) { instance_double(Logger, error: true) }
-      before { allow(Stenotype::Event).to receive(:new).and_raise(StandardError.new('boom')) }
 
-      context 'when graceful exception handling is enabled' do
+      before { allow(described_class).to receive(:new).and_raise(StandardError.new("boom")) }
+
+      context "when graceful exception handling is enabled" do
         before { Stenotype.configure { |config| config.logger = logger_double } }
 
-        it 'handles error' do
-          expect {
-            emit_event
-          }.to_not raise_error
+        it "handles error" do
+          expect { emit_event }.not_to raise_error
         end
       end
 
-      context 'when graceful exception handling is disabled' do
+      context "when graceful exception handling is disabled" do
         before do
           Stenotype.configure do |config|
             config.logger = logger_double
@@ -86,10 +85,8 @@ RSpec.describe Stenotype::Event do
 
         after { Stenotype.configure { |config| config.graceful_error_handling = true } }
 
-        it 'raises error' do
-          expect {
-            emit_event
-          }.to raise_error(Stenotype::Error)
+        it "raises error" do
+          expect { emit_event }.to raise_error(Stenotype::Error)
         end
       end
     end
@@ -125,21 +122,20 @@ RSpec.describe Stenotype::Event do
       end
     end
 
-    context 'when exception is raised' do
+    context "when exception is raised" do
       let(:logger_double) { instance_double(Logger, error: true) }
-      before { allow(test_dispatcher).to receive(:publish).and_raise(StandardError.new('boom')) }
 
-      context 'when graceful exception handling is enabled' do
+      before { allow(test_dispatcher).to receive(:publish).and_raise(StandardError.new("boom")) }
+
+      context "when graceful exception handling is enabled" do
         before { Stenotype.configure { |config| config.logger = logger_double } }
 
-        it 'handles error' do
-          expect {
-            event.emit!
-          }.to_not raise_error
+        it "handles error" do
+          expect { event.emit! }.not_to raise_error
         end
       end
 
-      context 'when graceful exception handling is disabled' do
+      context "when graceful exception handling is disabled" do
         before do
           Stenotype.configure do |config|
             config.logger = logger_double
@@ -149,10 +145,8 @@ RSpec.describe Stenotype::Event do
 
         after { Stenotype.configure { |config| config.graceful_error_handling = true } }
 
-        it 'raises error' do
-          expect {
-            event.emit!
-          }.to raise_error(Stenotype::Error)
+        it "raises error" do
+          expect { event.emit! }.to raise_error(Stenotype::Error)
         end
       end
     end